Add Vitest tests for parseResponse
